% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kfz}
\alias{kfz}
\title{1000 used cars in the USA crawled 2016}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
  \item{Marke}{brand}
  \item{PS}{horse power}
  \item{Preis}{price}
  \item{Alter}{age}
  \item{Kategorie}{category}
  \item{Kilometer}{mileage in km}
}
}
\usage{
kfz
}
\description{
The dataset was created from data available at \url{https://www.profheesen.de/}.
}
\keyword{datasets}
