% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.eval.wflset.r
\name{ml.eval.wflset}
\alias{ml.eval.wflset}
\title{Function to evaluate a machine learning Workflow set}
\usage{
ml.eval.wflset(modelle, daten)
}
\arguments{
\item{daten}{is a dataframe used for prediction}

\item{wflset}{is the name of a trained machine learning workflow set}
}
\value{
performance indicators rmse & rsq
}
\description{
Function to evaluate a machine learning Workflow set
}
\examples{
ml.eval.wflowset(mod.wflset.1,"autos.fold")
}
\keyword{evaluation}
\keyword{learning,}
\keyword{machine}
\keyword{modeling,}
