% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.korrelation.r
\name{ml.korrelation}
\alias{ml.korrelation}
\title{Function to calculate the correlation for numeric variables}
\usage{
ml.korrelation(df, titel, var)
}
\arguments{
\item{df}{is a dataframe with numeric variables}

\item{titel}{is the name of the variable to be used as subtitle on diagrams}

\item{var}{is a vector with the names of all variables in the dataframe df}
}
\value{
text with feedback to the function
}
\description{
Function to calculate the correlation for numeric variables
}
\examples{
ml.korrelation(mpg[,c(3,9,1)],"Autos",variablen3)
}
\keyword{correlation,}
\keyword{descriptive,}
\keyword{matrix}
\keyword{statistics,}
