% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.tutorial.R
\name{ml.tutorial}
\alias{ml.tutorial}
\title{Function to start a tutorial accompanying the package "pythonforbusiness" and the book "Künstliche Intelligenz im Business".}
\usage{
ml.tutorial(name)
}
\arguments{
\item{name}{is the name of the tutorial}
}
\description{
Function to start a tutorial accompanying the package "pythonforbusiness" and the book "Künstliche Intelligenz im Business".
}
\examples{
 ml.tutorial(name = "py.syntax")
}
\keyword{tutorial}
